+++ src/types/messages/index.ts
@@ -4,21 +4,7 @@
  * SPDX-License-Identifier: AGPL-3.0-only
  */
 
-import { Participant } from '../participant';
-import { SENSITIVITY_VALUES } from 'constants/index';
-import type { MailsEditorV2 } from 'types/editor/index.d';
-import type { PrefsType } from 'types/settings';
-import {
-	type MailVerificationHeader,
-	type MessageSignature,
-	type SaveDraftResponse,
-	type SoapMailMessagePart,
-	type SoapMailParticipant
-} from 'types/soap/index.d';
-
-type MailHeaderAttrs = {
-	[K in MailVerificationHeader]: K extends 'Authentication-Results' ? string | string[] : string;
-};
+import { MessageComputedSchema, MessageSchema } from '../schemas';
 
 export type MailAuthenticationHeader = { value: string; pass: boolean };
 
@@ -28,81 +14,23 @@ export type MailAuthenticationHeaders = {
 	dmarc?: MailAuthenticationHeader;
 };
 
-/**
- * Central catalog of all possible message properties.
- *
- * ⚠️ This type is NOT meant to be used directly as a runtime message shape.
- * It acts as a single source of truth for property types.
- *
- * All concrete message types (e.g. SearchMessage, FullMessage, DraftMessage)
- * must be derived from this schema
- *
- * If a property type changes, it should be updated here
- * so all derived types automatically stay in sync.
- *
- */
-export type MessageSchema = {
-	readonly id: string;
-	conversationId: string;
-	messageId: string;
-	parent: string;
-	size: number;
-	date: number;
-	sendDate: number;
-	revision: number;
-	flags: string;
-	originalMessageId: string;
-	tagNames: string;
-	tagIds: string;
-	replyType: 'r' | 'w';
-	subject: string;
-	fragment: string;
-	participants: Array<SoapMailParticipant>;
-	parts: Array<SoapMailMessagePart>;
-	autoSendTime: number;
-	invite: Array<any>;
-	shared: Array<any>;
-	signature: Array<MessageSignature>;
-	headers: Partial<MailHeaderAttrs>;
-};
+export type SoapMessageParticipant = MessageSchema['participants'][number];
+export type SoapMessagePart = MessageSchema['parts'][number];
+export type SoapMessageInvite = MessageSchema['invite'][number];
+export type SoapMessageShared = MessageSchema['shared'][number];
+export type SoapMessageSignature = MessageSchema['signature'][number];
 
-/**
- * Central catalog of all possible message computed properties.
- * Here are collected all the properties manipulated and normalized by the frontend.
- * If a new property is manipulated and its type or key change, it should be added here.
- */
-export type MessageComputedSchema = {
-	read: boolean;
-	hasAttachment: boolean;
-	flagged: boolean;
-	urgent: boolean;
-	isDeleted: boolean;
-	isSentByMe: boolean;
-	isForwarded: boolean;
-	isInvite: boolean;
-	isDraft: boolean;
-	isScheduled: boolean;
-	attachments: Array<AttachmentPart>;
-	tags: Array<string>;
-	parts: Array<MailMessagePart>;
-	participants: Array<Participant>;
-	body: BodyPart;
-	isComplete: boolean;
-	isReplied: boolean;
-	isReadReceiptRequested?: boolean;
-	isEncrypted: boolean;
-	messageIsFromExternalDomain?: boolean;
-	sensitivity: Sensitivity;
-	messageIdFromMailHeaders: string;
-	creationDateFromMailHeaders: string;
-	messageIsFromDistributionList: boolean;
-};
+export type MailMessageAttachment = MessageComputedSchema['attachments'][number];
+export type MailMessagePart = MessageComputedSchema['parts'][number];
+export type MailMessageParticipant = MessageComputedSchema['participants'][number];
+export type MailMessageSignature = MessageSchema['signature'][number];
+export type Sensitivity = MessageComputedSchema['sensitivity'];
 
 export type MailHeaders = {
 	signature?: MessageSchema['signature'];
 	messageIsFromExternalDomain?: MessageComputedSchema['messageIsFromExternalDomain'];
 	// authenticationHeaders: MailAuthenticationHeaders;
-	sensitivity?: MessageComputedSchema['sensitivity'];
+	sensitivity?: Sensitivity;
 	messageIdFromMailHeaders?: MessageComputedSchema['messageIdFromMailHeaders'];
 	creationDateFromMailHeaders?: MessageComputedSchema['creationDateFromMailHeaders'];
 	messageIsFromDistributionList?: MessageComputedSchema['messageIsFromDistributionList'];
@@ -143,38 +71,8 @@ export type IncompleteMessage = MailHeaders & {
 	isEncrypted?: MessageComputedSchema['isEncrypted'];
 };
 
-export type MailMessagePart = {
-	body?: boolean;
-	contentType: string;
-	size: number;
-	content?: string;
-	name: string;
-	filename?: string;
-	parts?: Array<MailMessagePart>;
-	ci?: string;
-	cd?: string;
-	disposition?: 'inline' | 'attachment';
-};
-
 export type MailMessagePartWithDisposition = MailMessagePart & {
 	disposition: 'inline' | 'attachment';
-};
-
-export type AttachmentPart = {
-	part?: string;
-	ct?: string;
-	s?: number;
-	size?: number;
-	filename?: string;
-	body?: boolean;
-	contentType?: string;
-	content?: string;
-	name?: string;
-	parts?: Array<AttachmentPart>;
-	ci?: string;
-	disposition?: 'inline' | 'attachment';
-	cd?: 'inline' | 'attachment';
-	mp?: Array<AttachmentPart>;
 };
 
 export type MailMessage = IncompleteMessage & {
@@ -183,24 +81,3 @@ export type MailMessage = IncompleteMessage & {
 	parent: MessageSchema['parent'];
 	isReadReceiptRequested?: MessageComputedSchema['isReadReceiptRequested'];
 };
-
-export type BodyPart = { contentType: string; content: string; truncated: boolean };
-/**
- * Parameters' type for the SendMsgRequest API command
- */
-export type SendMsgParameters = {
-	editor: MailsEditorV2;
-	msg?: MailMessage;
-	message?: MailMessage;
-	prefs?: PrefsType;
-};
-
-export type SendMsgResult = {
-	response:
-		| SaveDraftResponse
-		| (SaveDraftResponse['Fault'] & {
-				error: true;
-		  });
-};
-
-export type Sensitivity = (typeof SENSITIVITY_VALUES)[number];
