+++ src/views/app/detail-panel/preview/parts/info-block/smime-icon.tsx
@@ -8,10 +8,14 @@ import React from 'react';
 import { Icon, Row, Tooltip } from '@zextras/carbonio-design-system';
 import { useTranslation } from 'react-i18next';
 
-import { MessageSignature } from 'types/index.d';
+import { MailMessageSignature } from '../../../../../../types';
 import { getSignedIconColor } from 'views/app/detail-panel/preview/utils/index';
 
-export const SmimeIcon = ({ signature }: { signature: MessageSignature }): React.JSX.Element => {
+export const SmimeIcon = ({
+	signature
+}: {
+	signature: MailMessageSignature;
+}): React.JSX.Element => {
 	const [t] = useTranslation();
 	return (
 		<Tooltip
