+++ src/views/app/detail-panel/preview/parts/info-details-modal/mail-info-detail-modal.tsx
@@ -9,13 +9,13 @@ import { Container } from '@zextras/carbonio-design-system';
 import { ModalFooter, ModalHeader } from '@zextras/carbonio-ui-commons';
 import { useTranslation } from 'react-i18next';
 
-import { IncompleteMessage, MessageSignature } from 'types/index.d';
+import { IncompleteMessage, MailMessageSignature } from 'types/index.d';
 import { MailGeneralInfoSubsection } from 'views/app/detail-panel/preview/parts/info-details-modal/subsections/mail-general-info-subsection';
 import { SmimeSubsection } from 'views/app/detail-panel/preview/parts/info-details-modal/subsections/smime-subsection';
 
 type MailInfoDetailModalProps = {
 	onClose: () => void;
-	signature?: MessageSignature;
+	signature?: MailMessageSignature;
 	messageIdFromMailHeaders?: string;
 	creationDateFromMailHeaders?: string;
 	// authenticationMailsHeaders?: IncompleteMessage['authenticationHeaders'];
